requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
Pillow==9.5.0
PyPDF2==3.0.1
pandas==2.0.3
openpyxl==3.1.2
jinja2==3.1.2
gunicorn==21.2.0
flask-cors==4.0.0
anthropic==0.25.0
pillow-heif==0.13.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
numpy==1.24.3 